pyproject.toml
[project]
name = "bookmyseat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.120.3",
    "gradio>=5.23.1",
    "groq>=0.33.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "selenium>=4.38.0",
    "streamlit>=1.51.0",
    "streamlit-lottie>=0.0.5",
    "streamlit-mic-recorder>=0.0.8",
    "uvicorn>=0.38.0",
]
